Created a rule settor a README, and an area struct
The rule_settor class takes a filepath and is supposed to store all the
areas in the rules file. Its functionality is still incomplete. The
README.md describes the rules' syntax and the area struct is supposed to
store the area's name and color after being parsed by the rule settor
class.